requirements.txt
astroid==2.3.3
boto3==1.11.7
botocore==1.14.7
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
marshmallow==3.3.0
mccabe==0.6.1
oauthlib==3.1.0
pylint==2.4.4
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.1
simplejson==3.17.0
six==1.14.0
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
